Adding in transform plugins. CSVFormatter is added to begin with.
