+++ elasticsearch-plugins/src/test/java/co/cask/hydrator/plugin/testclasses/TableSink.java
@@ -1,0 +1,89 @@
+/*
+ * Copyright © 2015 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package co.cask.hydrator.plugin.testclasses;
+
+import co.cask.cdap.api.annotation.Description;
+import co.cask.cdap.api.annotation.Name;
+import co.cask.cdap.api.annotation.Plugin;
+import co.cask.cdap.api.data.format.StructuredRecord;
+import co.cask.cdap.api.data.schema.Schema;
+import co.cask.cdap.api.dataset.lib.KeyValue;
+import co.cask.cdap.api.dataset.table.Put;
+import co.cask.cdap.api.dataset.table.Table;
+import co.cask.cdap.etl.api.Emitter;
+import co.cask.cdap.etl.api.PipelineConfigurer;
+import co.cask.cdap.etl.api.batch.BatchRuntimeContext;
+import co.cask.cdap.etl.batch.sink.BatchWritableSink;
+import co.cask.cdap.etl.common.Properties;
+import co.cask.cdap.etl.common.RecordPutTransformer;
+import co.cask.cdap.etl.common.TableSinkConfig;
+import com.google.common.base.Preconditions;
+import com.google.common.base.Strings;
+import com.google.common.collect.Maps;
+
+import java.util.Map;
+
+/**
+ * CDAP Table Dataset Batch Sink.
+ */
+@Plugin(type = "batchsink")
+@Name("Table")
+@Description("Writes records to a Table with one record field mapping to the Table rowkey," +
+  " and all other record fields mapping to Table columns.")
+public class TableSink extends BatchWritableSink<StructuredRecord, byte[], Put> {
+
+  private final TableSinkConfig tableSinkConfig;
+  private RecordPutTransformer recordPutTransformer;
+
+  public TableSink(TableSinkConfig tableSinkConfig) {
+    this.tableSinkConfig = tableSinkConfig;
+  }
+
+  @Override
+  public void configurePipeline(PipelineConfigurer pipelineConfigurer) {
+    super.configurePipeline(pipelineConfigurer);
+    Preconditions.checkArgument(!Strings.isNullOrEmpty(tableSinkConfig.getRowField()),
+                                "Row field must be given as a property.");
+  }
+
+  @Override
+  public void initialize(BatchRuntimeContext context) throws Exception {
+    super.initialize(context);
+    Schema outputSchema = null;
+    // If a schema string is present in the properties, use that to construct the outputSchema and pass it to the
+    // recordPutTransformer
+    String schemaString = context.getPluginProperties().getProperties().get(Properties.Table.PROPERTY_SCHEMA);
+    if (schemaString != null) {
+      outputSchema = Schema.parseJson(schemaString);
+    }
+    recordPutTransformer = new RecordPutTransformer(tableSinkConfig.getRowField(), outputSchema);
+  }
+
+  @Override
+  protected Map<String, String> getProperties() {
+    Map<String, String> properties = Maps.newHashMap(tableSinkConfig.getProperties().getProperties());
+    properties.put(Properties.BatchReadableWritable.NAME, tableSinkConfig.getName());
+    properties.put(Properties.BatchReadableWritable.TYPE, Table.class.getName());
+    return properties;
+  }
+
+  @Override
+  public void transform(StructuredRecord input, Emitter<KeyValue<byte[], Put>> emitter) throws Exception {
+    Put put = recordPutTransformer.toPut(input);
+    emitter.emit(new KeyValue<>(put.getRow(), put));
+  }
+}
