+++ transform-plugins/resources/plugin/hydrator-transform-plugins.json
@@ -1,0 +1,3 @@
+{
+  "parents" : [ "system:cdap-etl-batch[3.2.0,3.3.0)", "system:cdap-etl-realtime[3.2.0,3.3.0)" ]
+}
