+++ transform-plugins/resources/ui/CSVFormatter.json
@@ -1,0 +1,36 @@
+{
+  "id": "CSVFormatter",
+  "groups": {
+    "position": [ "group1" ],
+    "group1": {
+      "display": "CSV Formatter Configuration",
+      "position": [ "format", "delimiter", "schema" ],
+      "fields": {
+        "delimiter": {
+          "widget": "select",
+          "label": "Choose a delimiter",
+          "properties": {
+            "values" : [ "COMMA", "CTRL-A", "TAB", "VBAR", "STAR", "CARROT", "DOLLAR", "HASH", "CTRL-B", "CTRL-C", 
+              "CTRL-D", "CTRL-E", "CTRL-F" ],
+            "default": "CTRL-A"
+          }
+        },
+        "format": {
+          "widget": "select",
+          "label": "Formats",
+          "properties": {
+            "values" : [ "DELIMITED", "EXCEL", "MYSQL", "RFC4180", "TDF" ],
+            "default": "DELIMITED"
+          }
+        },
+        "schema" : {
+          "widget": "schema",
+          "label": "Schema",
+          "description" : "Output Schema",
+          "schema-types" : [ "string" ],
+          "schema-default-type" : "string"
+        }
+      }
+    }
+  }
+}
