+++ transform-plugins/src/main/java/co/cask/hydrator/plugin/CSVFormatter.java
@@ -1,0 +1,7 @@
+package co.cask.hydrator.plugin;
+
+/**
+ * Created by nmotgi on 10/25/15.
+ */
+public class CSVFormatter {
+}
