+++ transform-plugins/src/test/java/co/cask/hydrator/plugin/CSVFormatterTest.java
@@ -1,0 +1,10 @@
+package co.cask.hydrator.plugin;
+
+import static org.junit.Assert.*;
+
+/**
+ * Created by nmotgi on 10/25/15.
+ */
+public class CSVFormatterTest {
+
+}
